Renames ap-heatmap to attack-patterns-heatmap, since we just replaced the intrusion-sets component by the same name
