+++ src/app/global/components/heatmap/attack-patterns-heatmap.component.html
