+++ src/app/global/global.module.ts
@@ -62,7 +62,7 @@ import { RelatationshipGeneratorComponent } from './components/relatationship-ge
 import { InfiniteScrollDirective } from './directives/infinite-scroll.directive';
 import { PublishedCheckboxComponent } from './components/published-checkbox/published-checkbox.component';
 import { HeaderLogoComponent } from './components/header-logo/header-logo.component';
-import { ApHeatmapComponent } from './components/heatmap/ap-heatmap.component';
+import { AttackPatternsHeatmapComponent } from './components/heatmap/attack-patterns-heatmap.component';
 
 const matModules = [
     MatAutocompleteModule,
@@ -117,7 +117,7 @@ const unfetterComponents = [
     RiskBreakdownComponent,
     TreemapComponent,
     HeatmapComponent,
-    ApHeatmapComponent,
+    AttackPatternsHeatmapComponent,
     CreatedByRefComponent,
     RelatationshipGeneratorComponent,
     InfiniteScrollDirective,
